Tarefa1_2022li1g008

Description

Módulo para a realização da Tarefa 1 do projeto de LI1 em 2022/23.

Documentation

mapaValido

Arguments

:: Mapa
É escrito um tipo de mapa qualquer.

-> Bool
A função irá avaliar se o mapa inserido é válido (devolvendo True) ou se não o é (devolvendo False).

A função mapaValido verifica se um dado mapa obedece a todas as regras presentes na Tarefa 1, utilizando para isso algumas funções auxiliares.

Exemplos de utilização:

>>> mapaValido (Mapa 5 [(Rio 1, [Tronco, Nenhum, Tronco, Nenhum, Tronco]),
                        (Rio (-1), [Nenhum, Nenhum, Nenhum, Tronco, Tronco]),
                        (Estrada 2, [Nenhum, Carro, Carro, Carro, Nenhum])])
True
>>> mapaValido (Mapa 5 [(Rio 1, [Tronco, Nenhum, Tronco, Nenhum, Tronco]),
                        (Rio 1, [Nenhum, Nenhum, Nenhum, Tronco, Tronco]),
                        (Estrada 2, [Nenhum, Carro, Carro, Carro, Nenhum])])
False
>>> mapaValido (Mapa 5 [(Rio 1, [Tronco, Nenhum, Arvore, Nenhum, Tronco]),
                        (Rio (-1), [Nenhum, Nenhum, Nenhum, Tronco, Tronco]),
                        (Estrada 2, [Nenhum, Carro, Carro, Carro, Nenhum])])
False

velocity

Arguments

:: [(Terreno, [Obstaculo])]
A função recebe uma lista de pares, no qual o primeiro elemento de cada par é um terreno, e o segundo elemento é uma lista de obstáculos desse mesmo terreno.

-> Bool
A função retornará valor lógico False caso a velocidade da Relva seja diferente de 0, e a velocidade de um Rio e de uma Estrada seja 0.

A função velocity serve para indicar se a velocidade do terreno é possível.

getvelocidade

Arguments

:: Terreno
A função recebe um tipo de terreno com a sua velocidade correspondente.

-> Int
A função retorna a velocidade correspondente ao terreno que ela recebeu.

A função getvelocidade serve para indicar qual a velocidade correspondente ao tipo de terreno.

obstaculo

Arguments

:: [(Terreno, [Obstaculo])]
Escreve-se uma lista de pares em que o primeiro elemento do par é o tipo de terreno e o segundo elemento é a lista de obstáculos presentes nesse terreno.

-> Bool
Caso não existam obstáculos impróprios para cada tipo de terreno associado, o resultado será True.

A função obstaculo verifica se não existem obstáculos em terrenos imprórios, por exemplo não podem existir carros ou árvores nos rios, nem troncos ou árvores nas estradas, e nem carros ou troncos na relva.

dirRios

Arguments

:: [(Terreno, [Obstaculo])]
Escreve-se uma lista de pares em que o primeiro elemento do par é o tipo de terreno e o segundo elemento é a lista de obstáculos presentes nesse terreno.

-> Bool
Caso dois rios seguidos tenham valores de velocidade opostos, então o resultado será True.

A função dirRios verifica se rios contíguos/seguidos têm direções opostas, ou seja, se as velocidades de dois rios consecutivos têm sinais opostos.

troncos

Arguments

:: [(Terreno, [Obstaculo])]
Escreve-se uma lista de pares em que o primeiro elemento do par é o tipo de terreno e o segundo elemento é a lista de obstáculos presentes nesse terreno.

-> Bool
Caso um tronco tenha mais do que 5 unidades de comprimento, ou seja, caso a palavra Tronco apareça mais do que 5 vezes seguidas na lista,o resultado será False.

A função troncos tem como papel principal assegurar-se de que os troncos têm no máximo 5 unidades de comprimento.

troncosaux1

Arguments

:: Int
Escreve-se o número 0.

-> (Terreno, [Obstaculo])
Escreve-se um par composto por um tipo de terreno e por uma lista de obstáculos desse mesmo terreno.

-> Int
A função revela qual o comprimento de de um tronco que esteja exatamente no início do rio.

A função troncosaux1 é uma das funções auxiliares da função "troncos" para que juntamente com as outras funções auxiliares façam com que a função "troncos" retorne que uma lista de obstáculos é inválida caso os obstáculos dessa lista ao sairem do mapa e voltando a reaparecer do lado oposto não originem um tronco com mais do que 5 unidades de comprimento.

troncosaux2

Arguments

:: Int
Escreve-se o número 0.

-> (Terreno, [Obstaculo])
Escreve-se um par composto por um tipo de terreno e por uma lista de obstáculos desse mesmo terreno.

-> Int
A função revela qual o comprimento de de um tronco que esteja exatamente no final do rio.

A função troncosaux2 é mais uma das funções auxiliares da função "troncos".

troncosaux3

Arguments

:: (Terreno, [Obstaculo])
Escreve-se um par composto por um tipo de terreno e por uma lista de obstáculos desse mesmo terreno.

-> Int
A função revela qual o comprimento de um tronco que tem uma parte na extremidade esquerda do rio e a outra parte na extremidade direita do rio.

A função troncosaux3 é mais uma das funções auxiliares da função "troncos".

troncosaux4

Arguments

:: Int
Escreve-se o número 0.

-> (Terreno, [Obstaculo])
Escreve-se um par composto por um tipo de terreno e por uma lista de obstáculos desse mesmo terreno.

-> Bool
Se um tronco tiver mais do que 5 unidades de comprimento, esta função retornará o resultado False.

A função troncosaux4 é mais uma das funções auxiliares da função "troncos".

carros

Arguments

:: [(Terreno, [Obstaculo])]
Escreve-se uma lista de pares em que o primeiro elemento do par é o tipo de terreno e o segundo elemento é a lista de obstáculos presentes nesse terreno.

-> Bool
Caso um carro tenha mais do que 3 unidades de comprimento, ou seja, caso a palavra Carro apareça mais do que 3 vezes seguidas na lista, o resultado será False.

A função carros serve para limitar o comprimento dos carros no máximo a 3 unidades de comprimento.

carrosaux1

Arguments

:: Int
Escreve-se o número 0.

-> (Terreno, [Obstaculo])
Escreve-se um par composto por um tipo de terreno e por uma lista de obstáculos desse mesmo terreno.

-> Int
A função revela qual o comprimento de de um carro que esteja exatamente no início da estrada.

A função carrosaux1 é uma das funções auxiliares da função "carros" para que juntamente com as outras funções auxiliares façam com que a função "carros" retorne que uma lista de obstáculos é inválida caso os obstáculos dessa lista ao sairem do mapa e voltando a reaparecer do lado oposto não originem um carro com mais do que 3 unidades de comprimento.

carrosaux2

Arguments

:: Int
Escreve-se o número 0.

-> (Terreno, [Obstaculo])
Escreve-se um par composto por um tipo de terreno e por uma lista de obstáculos desse mesmo terreno.

-> Int
A função revela qual o comprimento de de um carro que esteja exatamente no final da estrada.

A função carrosaux2 é mais uma das funções auxiliares da função "carros".

carrosaux3

Arguments

:: (Terreno, [Obstaculo])
Escreve-se um par composto por um tipo de terreno e por uma lista de obstáculos desse mesmo terreno.

-> Int
A função revela qual o comprimento de um carro que tem uma parte na extremidade esquerda da estrada e a outra parte na extremidade direita da estrada.

A função carrosaux3 é mais uma das funções auxiliares da função "carros".

carrosaux4

Arguments

:: Int
Escreve-se o número 0.

-> (Terreno, [Obstaculo])
Escreve-se um par composto por um tipo de terreno e por uma lista de obstáculos desse mesmo terreno.

-> Bool
Se um carro tiver mais do que 3 unidades de comprimento, esta função retornará o resultado False.

A função carrosaux4 é mais uma das funções auxiliares da função "carros".

nenhuns

Arguments

:: [(Terreno, [Obstaculo])]
Escreve-se uma lista de pares em que o primeiro elemento do par é o tipo de terreno e o segundo elemento é a lista de obstáculos presentes nesse terreno.

-> Bool
Caso exista pelo menos um obstáculo Nenhum na lista de obstáculos para cada tipo de terreno associado, o resultado será True.

A função nenhuns verifica se existe no mínimo um obstáculo Nenhum, ou seja, uma linha não pode ser composta exclusivamente por obstáculos, precisando de haver pelo menos um espaço livre

largura

Arguments

:: Int

-> [(Terreno, [Obstaculo])]
Escreve-se uma lista de pares em que o primeiro elemento do par é o tipo de terreno e o segundo elemento é a lista de obstáculos presentes nesse terreno.

-> Bool
Caso o número correspondente à largura do mapa seja igual ao comprimento da lista de obstáculos para cada tipo de terreno associado, o resultado será True.

A função largura verifica se o número de obstáculos de uma lista corresponde à largura do mapa

rios

Arguments

:: Int
Escreve-se o número 0 que servirá para contabilizar o número de rios que aparecem seguidos, e cada vez que o número de rios seguidos for inferior ou igual a 4 e a seguir estiver escrito outro tipo de terreno, este número voltará a 0.

-> [(Terreno, [Obstaculo])]
Escreve-se uma lista de pares em que o primeiro elemento do par é o tipo de terreno e o segundo elemento é a lista de obstáculos presentes nesse terreno.

-> Bool
Caso se encontrem mais do que 4 rios consecutivos, ou seja, caso a palavra Rio apareça mais do que 4 vezes em pares consecutivos, o resultado será False.

A função rios verifica se não existem mais do que 4 rios consecutivos

estradas

Arguments

:: Int
Escreve-se o número 0 que servirá para contabilizar o número de estradas que aparecem seguidas, e cada vez que o número de estradas seguidas for inferior ou igual a 5 e a seguir estiver escrito outro tipo de terreno, este número voltará a 0.

-> [(Terreno, [Obstaculo])]
Escreve-se uma lista de pares em que o primeiro elemento do par é o tipo de terreno e o segundo elemento é a lista de obstáculos presentes nesse terreno.

-> Bool
Caso se encontrem mais do que 5 estradas consecutivas, ou seja, caso a palavra Estrada apareça mais do que 5 vezes em pares consecutivos, o resultado será False.

A função estradas verifica se não existem mais do que 5 estradas consecutivas

relvas

Arguments

:: Int
Escreve-se o número 0 que servirá para contabilizar o número de relvas que aparecem seguidas, e cada vez que o número de relvas seguidas for inferior ou igual a 5 e a seguir estiver escrito outro tipo de terreno, este número voltará a 0.

-> [(Terreno, [Obstaculo])]
Escreve-se uma lista de pares em que o primeiro elemento do par é o tipo de terreno e o segundo elemento é a lista de obstáculos presentes nesse terreno.

-> Bool
Caso se encontrem mais do que 5 relvas consecutivas, ou seja, caso a palavra Relva apareça mais do que 5 vezes em pares consecutivos, o resultado será False.

A função relvas verifica se não existem mais do que 5 relvas consecutivas