Tarefa2_2022li1g008

Description

Módulo para a realização da Tarefa 2 do projeto de LI1 em 2022/23.

Documentation

estendeMapa

Arguments

:: Mapa
A função recebe um mapa válido.

-> Int
A função recebe também um número inteiro entre 0 e 100 para gerar aleatoriamente uma nova linha.

-> Mapa
A função gera um novo mapa válido com a nova linha no topo.

A função estendeMapa dá-nos um novo mapa válido com uma nova linha gerada aleatoriamente.

Exemplos de utilização:

>>> estendeMapa (Mapa 4 [(Relva,[Nenhum,Arvore,Arvore,Nenhum]),
                         (Rio (-3),[Tronco,Tronco,Nenhum,Tronco]),
                         (Rio 1,[Nenhum,Nenhum,Tronco,Nenhum]),
                         (Estrada 2,[Carro,Carro,Carro,Nenhum])]) 0
Mapa 4 [(Rio 5,[Nenhum,Nenhum,Nenhum,Tronco]),
        (Relva,[Nenhum,Arvore,Arvore,Nenhum]),
        (Rio (-3),[Tronco,Tronco,Nenhum,Tronco]),
        (Rio 1,[Nenhum,Nenhum,Tronco,Nenhum]),
        (Estrada 2,[Carro,Carro,Carro,Nenhum])] 

>>> estendeMapa (Mapa 6 [(Relva,[Arvore,Arvore,Arvore,Nenhum,Nenhum,Arvore]),
                         (Relva,[Nenhum,Arvore,Nenhum,Nenhum,Arvore,Arvore]),
                         (Relva,[Nenhum,Nenhum,Nenhum,Nenhum,Arvore,Arvore]),
                         (Relva,[Arvore,Arvore,Arvore,Nenhum,Arvore,Nenhum]),
                         (Relva,[Arvore,Nenhum,Nenhum,Nenhum,Arvore,Arvore])]) 95
Mapa 6 [(Estrada 1,[Carro,Carro,Nenhum,Carro,Nenhum,Carro]),
        (Relva,[Arvore,Arvore,Arvore,Nenhum,Nenhum,Arvore]),
        (Relva,[Nenhum,Arvore,Nenhum,Nenhum,Arvore,Arvore]),
        (Relva,[Nenhum,Nenhum,Nenhum,Nenhum,Arvore,Arvore]),
        (Relva,[Arvore,Arvore,Arvore,Nenhum,Arvore,Nenhum]),
        (Relva,[Arvore,Nenhum,Nenhum,Nenhum,Arvore,Arvore])] 

>>> estendeMapa (Mapa 6 [(Rio 1,[Tronco,Tronco,Tronco,Nenhum,Tronco,Nenhum]),
                         (Rio (-2),[Nenhum,Tronco,Tronco,Nenhum,Tronco,Tronco]),
                         (Rio 4,[Tronco,Nenhum,Tronco,Tronco,Nenhum,Tronco]),
                         (Rio (-4),[Tronco,Tronco,Tronco,Nenhum,Nenhum,Tronco])]) 23
Mapa 6 [(Estrada (-1),[Nenhum,Nenhum,Nenhum,Carro,Carro,Carro]),
        (Rio 1,[Tronco,Tronco,Tronco,Nenhum,Tronco,Nenhum]),
        (Rio (-2),[Nenhum,Tronco,Tronco,Nenhum,Tronco,Tronco]),
        (Rio 4,[Tronco,Nenhum,Tronco,Tronco,Nenhum,Tronco]),
        (Rio (-4),[Tronco,Tronco,Tronco,Nenhum,Nenhum,Tronco])] 

estendeLinha

Arguments

:: Int
A função recebe um número inteiro a partir do qual inicia a aleatoriedade.

-> Int
A função recebe também um número inteiro correspondente ao comprimento da lista de obstáculos.

-> Velocidade
A função recebe também uma velocidade.

-> (Terreno, [Obstaculo])
A função recebe um par, com um terreno como primeiro elemento e uma lista de obstáculos como segundo elemento.

-> (Terreno, [Obstaculo])
A função gera um novo par (nova linha do mapa) aleatório.

A função estendeLinha dá-nos uma nova linha gerada aleatoriamente.

randomList :: Random a => Int -> [a]

A função randomList gera uma lista aleatória.

randomInt :: Int -> Int -> Int

A função randomInt gera um número aleatório.

geraAleatorios

Arguments

:: Int
A função recebe um número inteiro que limita o intervalo de aleatoriedade.

-> Int
A função recebe um número inteiro que indica o número de inteiros que queremos para fazer essa mesma aleatoriedade.

-> Int
A função recebe um número inteiro a partir do qual inicia a aleatoriedade.

-> [Int]
A função gera uma lista de números inteiros.

A função geraAleatorios dá-nos uma lista de números escolhida aleatoriamente.

geraObsFromAleatorios

Arguments

:: [Int]
A função recebe uma lista de números inteiros.

-> [Obstaculo]
A função recebe uma lista de obstáculos válida.

-> [Obstaculo]
A função gera uma nova lista de obstáculos com base nos obstáculos válidos inseridos.

A função geraObsFromAleatorios dá-nos uma lista de obstáculos escolhida aleatoriamente com base nos obstáculos válidos.

geraTerrenoFromAleatorios

Arguments

:: [Int]
A função recebe uma lista com um número inteiro.

-> [Terreno]
A função recebe também uma lista com um Terreno.

-> Terreno
A função retorna esse mesmo Terreno.

A função geraTerrenoFromAleatorios dá-nos um Terreno escolhido aleatoriamente.

proximosTerrenosValidos

Arguments

:: Mapa
A função recebe um mapa válido.

-> [Terreno]
A função retorna-nos uma lista de terrenos possíveis de inserir na seguinte linha.

A função proximosTerrenosValidos indica os terrenos possíveis de inserir na seguinte linha de um mapa dado.

proximosObstaculosValidos

Arguments

:: Int
A função recebe um número inteiro correspondente à largura do mapa.

-> (Terreno, [Obstaculo])
A função recebe também um par com um determinado terreno e a respetiva lista de obstáculos.

-> [Obstaculo]

A função proximosObstaculosValidos indica os obstáculos possíveis de inserir numa lista de obstáculos dada.

listaObstaculos

Arguments

:: [Obstaculo]
A função recebe uma lista de obstáculos.

-> Bool
A função retorna o valor lógico True se não existirem mais do que 5 troncos consecutivos nem mais do que 3 carros consecutivos.

A função listaObstaculos verifica se não existem mais do que 5 troncos consecutivos nem mais do que 3 carros consecutivos.

tamanhoObstaculos

Arguments

:: [[Obstaculo]]
A função recebe uma lista de listas de obstáculos.

-> Bool
Caso o primeiro elemento de uma lista seja Tronco ou Carro e o comprimento da lista seja maior que 5 ou que 3, respetivamente, o resultado será False.

A função tamanhoObstaculos verifica se não existem listas de comprimento superior a 5 e com um Tronco como primeiro elemento dessa lista, e se não existem listas de comprimento superior a 3 e com um Carro como primeiro elemento dessa lista.