Tarefa3_2022li1g008

Description

Módulo para a realização da Tarefa 3 do projeto de LI1 em 2022/23.

Documentation

animaJogo

Arguments

:: Jogo
A função recebe um jogo, ou seja, as coordenadas iniciais do jogador e o mapa onde este se encontra.

-> Jogada
E recebe também a jogada que o jogador irá descrever.

-> Jogo
A função retornará a nova posição do jogador consoante a jogada que este efetuou, e o novo mapa após as obstáculos se terem movido de acordo com a velocidade do terreno onde estão inseridos.

A função animaJogo movimenta os obstáculos do terreno tendo em conta a velocidade do mesmo, e o personagem, de acordo com a jogada dada.

Exemplos de utilização:

>>> animaJogo (Jogo (Jogador (2,1))
                (Mapa 3 [(Rio 2,[Tronco,Tronco,Nenhum]),
                (Relva,[Arvore,Arvore,Nenhum]),
                (Estrada 1,[Nenhum,Nenhum,Carro])])) 
                (Move Esquerda)
Jogo (Jogador (2,1)) 
(Mapa 3 [(Rio 2,[Tronco,Nenhum,Tronco]),
(Relva,[Arvore,Arvore,Nenhum]),
(Estrada 1,[Carro,Nenhum,Nenhum])])

>>> animaJogo (Jogo (Jogador (0,2))
                (Mapa 3 [(Rio 2,[Tronco,Tronco,Nenhum]),
                (Relva,[Arvore,Arvore,Nenhum]),
                (Estrada 2,[Nenhum,Nenhum,Carro])])) 
                (Move Baixo)
Jogo (Jogador (0,2)) 
(Mapa 3 [(Rio 2,[Tronco,Nenhum,Tronco]),
(Relva,[Arvore,Arvore,Nenhum]),
(Estrada 2,[Carro,Nenhum,Nenhum])])

>>> animaJogo (Jogo (Jogador (0,0))
                (Mapa 3 [(Rio 2,[Tronco,Tronco,Nenhum]),
                (Relva,[Arvore,Arvore,Nenhum]),
                (Estrada 1,[Nenhum,Nenhum,Carro])])) 
                (Parado)
Jogo (Jogador (2,0)) 
(Mapa 3 [(Rio 2,[Tronco,Nenhum,Tronco]),
(Relva,[Arvore,Arvore,Nenhum]),
(Estrada 1,[Carro,Nenhum,Nenhum])])

mover

Arguments

:: Jogador
A função recebe a posição em que o jogador se encontra inicialmente.

-> Mapa
A função recebe um tipo de mapa válido.

-> Mapa
A função retorna o mesmo mapa mas com os obstáculos de cada terreno nas suas novas posições, uma vez que é a velociade do terreno que faz com que essas posições variem.

A função mover faz com que cada obstáculo da lista de obstáculos de cada terreno vá mudando de posição de acordo com a velocidade desse mesmo terreno.

mover2

Arguments

:: Jogador
A função recebe a posição em que o jogador se encontra inicialmente.

-> [(Terreno, [Obstaculo])]
A função recebe uma lista de pares, cujo primeiro elemento de cada par é um tipo de terreno e o segundo elemento de cada par é a lista de obstáculos desse mesmo terreno.

-> [(Terreno, [Obstaculo])]
A função retorna a mesma lista de pares, mas com os obstáculos de cada terreno nas suas novas posições, uma vez que é a velocidade do terreno que faz com que essas posições variem.

A função mover2 tem exatamente o mesmo objetivo do que a função "mover" (sendo portanto uma função auxiliar da mesma), com a particularidade de que não é necessário escrever a palavra Mapa e nem a sua largura.

mover3

Arguments

:: Int
Escreve-se o número 0 que irá contabilizar o número de vezes que a lista de obstáculos se move de cada vez até que este número seja igual ao da velocidade do terreno.

-> Jogador
A função recebe a posição em que o jogador se encontra inicialmente.

-> [(Terreno, [Obstaculo])]
A função recebe uma lista de pares, cujo primeiro elemento de cada par é um tipo de terreno e o segundo elemento de cada par é a lista de obstáculos desse mesmo terreno.

-> [(Terreno, [Obstaculo])]
A função retorna a mesma lista de pares, mas com os obstáculos de cada terreno nas suas novas posições, uma vez que é a velocidade do terreno que faz com que essas posições variem.

A função mover3 tem exatamente o mesmo objetivo do que a função "mover2", mas é uma função auxiliar da mesma.

mover4

Arguments

:: Int
Escreve-se o número 0 que irá contabilizar o número de vezes que a lista de obstáculos se move de cada vez até que este número seja igual ao da velocidade do terreno.

-> [(Terreno, [Obstaculo])]
A função recebe uma lista de pares, cujo primeiro elemento de cada par é um tipo de terreno e o segundo elemento de cada par é a lista de obstáculos desse mesmo terreno.

-> [(Terreno, [Obstaculo])]
A função retorna a mesma lista de pares, mas com os obstáculos de cada terreno nas suas novas posições, uma vez que é a velocidade do terreno que faz com que essas posições variem.

A função mover4 é uma função auxiliar da "mover3".

posicao

Arguments

:: Mapa
A função recebe um mapa válido.

-> Jogador
A função recebe a posição em que o jogador se encontra inicialmente.

-> [Jogada]
E recebe também um conjunto de jogadas que o jogador descreverá.

-> Jogador
A função retornará a posição onde o jogador se encontra após ter descrito determinadas jogadas.

A função posicao serve para indicar qual será a nova posição do jogador após algumas jogadas.

segueTronco

Arguments

:: Mapa
A função recebe um mapa válido.

-> Jogador
Recebe também as coordenadas de onde o jogador se situa inicialmente (salientando que esta função apenas será chamada caso o jogador se situe em cima de um tronco).

-> Jogada
E a jogada efetuada pelo jogador, que neste caso como ele está parado em cima do tronco, a jogada será Parado.

-> Jogador
A função retornará as coordenadas de onde o jogador se encontra após ser arrastado pelo tronco.

A função segueTronco serve para dizer as posições em que o jogador se vai situar caso este esteja parado em cima de um tronco.

encontrax

Arguments

:: Int
A função recebe um número que irá corresponder à posição de um elemento na lista.

-> [a]
E recebe uma lista.

-> a
E devolve o elemento dessa lista que se situa na posição na posição inserida

A função encontrax devolve um elemento de uma lista que se encontra na posição escolhida.

getListaObs

Arguments

:: (Terreno, [Obstaculo])
A função recebe uma linha do mapa, ou seja, um par onde o primeiro elemento é o tipo de terreno e o segundo elemento é a lista de obstáculos presentes nesse terreno.

-> [Obstaculo]
A função retorna a lista de obstáculos do par que recebeu inicialmente.

A função getListaObs serve para referir a lista de obstáculos presente numa linha do mapa.

getTerreno

Arguments

:: (Terreno, [Obstaculo])
A função recebe uma linha do mapa, ou seja, um par onde o primeiro elemento é o tipo de terreno e o segundo elemento é a lista de obstáculos presentes nesse terreno.

-> Terreno
A função retorna o tipo de terreno do par que recebeu inicialmente.

A função getTerreno serve para referir o tipo de terreno presente numa linha do mapa.

getMapa

Arguments

:: Mapa
A função recebe um tipo de mapa válido.

-> [(Terreno, [Obstaculo])]
E devolve todas as linhas desse mesmo mapa.

A função getMapa serve para referir as linhas presentes num mapa.

encontraCoordenadas

Arguments

:: Coordenadas
A função recebe um par de coordenadas.

-> Mapa
E recebe também um mapa válido.

-> Obstaculo
A função devolve qual o obstáculo do mapa que se encontra nas coordenadas que a função recebeu.

A função encontraCoordenadas serve para indicar qual o obstáculo presente nas coordenadas especificadas.

encontraCoordenadas2

Arguments

:: Coordenadas
A função recebe um par de coordenadas.

-> [(Terreno, [Obstaculo])]
E recebe também as linhas de um mapa válido.

-> Obstaculo
A função devolve qual o obstáculo do mapa que se encontra nas coordenadas que a função recebeu.

A função encontraCoordenadas2 desempenha extamente o mesmo papel que a função "encontraCoordenadas" só que em vez de receber um mapa recebe as linhas de um mapa.

getvelocidade

Arguments

:: Terreno
A função recebe um tipo de terreno com a sua velocidade correspondente.

-> Int
A função retorna a velocidade correspondente ao terreno que ela recebeu.

A função getvelocidade serve para indicar qual a velocidade correspondente ao tipo de terreno.