Tarefa6_2022li1g008

Description

Módulo para a realização da Tarefa 6 do projeto de LI1 em 2022/23.

Documentation

data Opcao

Constructors

Jogar

NovoJogo

Pausa

Sair

Abandonar

data Menu

Constructors

MenuInicial Opcao

ModoJogo

ModoBot

ModoBotJogo

PerdeuJogo

type World = (Menu, Jogo, Imagens, Tempo, TempoLinha, Pontuação)

type TempoLinha = Float

type Tempo = Float

type Imagens = [Picture]

type Pontuação = Int

jogoInicial

Arguments

:: Jogo
A função devolve um Jogo.

A função jogoInicial representa qual é o mapa inicial e quais as coordenadas iniciais do jogador.

window :: Display

A função window é simplesmente o tipo de janela onde o jogo será executado.

fr :: Int

A função fr corresponde ao frame rate.

drawState

Arguments

:: World
A função utiliza um valor do tipo World para saber qual o estado atual do jogo.

-> Picture
E retorna um valor do tipo Picture consoante o estado do jogo.

A função drawState converte o estado atual do jogo num valor do tipo Picture.

drawOption

Arguments

:: String
A função recebe uma string.

-> Picture
E depois transforma-a num valor do tipo Picture.

A função drawOption é uma função auxiliar da "drawState" que serve para ajudar a posicionar a String indicada.

desenhaMapa

Arguments

:: Mapa
A função recebe um mapa.

-> (Float, Float)
Recebe também um par de Floats que inicialmente será (0,0) e que serve para a função saber em que coordenadas irá desenhar cada obstáculo.

-> Imagens
E recebe ainda o conjunto de todas as imagens possíveis de serem utilizadas.

-> [Picture]
A função retorna uma lista de Pictures.

A função desenhaMapa tem como objetivo transformar o mapa do jogo num valor do tipo Picture.

desenhaLinha

Arguments

:: (Terreno, [Obstaculo])
A função recebe uma linha do mapa.

-> (Float, Float)
Recebe também um par de Floats que inicialmente será (0,0) e que serve para a função saber em que coordenadas irá desenhar cada obstáculo.

-> Imagens
E recebe ainda o conjunto de todas as imagens possíveis de serem utilizadas.

-> [Picture]
A função retorna uma lista de Pictures.

A função desenhaLinha é uma função auxiliar da "desenhaMapa" e tem como objetivo transformar uma linha do mapa do jogo num valor do tipo Picture.

desenhaObsTerr

Arguments

:: (Terreno, Obstaculo)
A função recebe o tipo de terreno com o respetivo obstáculo a desenhar (se este for Nenhum, a função apenas irá desenhar o tipo de terreno).

-> (Float, Float)
Recebe também um par de Floats que inicialmente será (0,0) e que serve para a função saber em que coordenadas irá desenhar cada obstáculo.

-> Imagens
E recebe ainda o conjunto de todas as imagens possíveis de serem utilizadas.

-> Picture
A função retorna os obstáculos e os terrenos sobre a forma do tipo Picture.

A função desenhaObsTerr é uma função auxiliar da "desenhaLinha" e tem como objetivo transformar os obstáculos e os terrenos do jogo num valor do tipo Picture.

event

Arguments

:: Event
A função recebe um evento (neste caso quando teclas em específico são pressionadas).

-> World
E recebe um valor do tipo World para saber qual o estado atual do jogo.

-> World
A função retorna o novo estado do jogo através do tipo World.

A função event calcula o próximo estado após reagir a um evento.

deslizaJogo2

Arguments

:: Int
A função recebe um número inteiro para gerar aleatoriamente uma nova linha.

-> Jogo
A função recebe um jogo (a posição do jogador e um mapa válido).

-> Jogo
A função retornará um novo jogo, com as mesmas coordenadas que o jogador tinha anteriormente mas agora com uma unidade somada ao "y", e o mesmo mapa inicialmente inserido mas com uma nova linha no topo (com as novas propriedade agora definidas) e com a última linha removida.

A função deslizaJogo2 faz exatamente o mesmo que a função "deslizaJogo", mas não permite que sejam geradas linhas cuja velocidade seja superior a 3 e nem relvas com mais do que 4 árvores e nem estradas com mais de 3 carros, e apenas é chamada quando a pontuação do jogador for inferior ou igual a 200, ou seja serve para que a partir desse valor a dificuldade do jogo aumente.

novoMapa

Arguments

:: Int
A função recebe um número inteiro para gerar aleatoriamente uma nova linha.

-> Mapa
A função recebe também um mapa válido.

-> Mapa
A função gera um novo mapa válido com a nova linha no topo (cuja velocidade tem de ser inferior a 3).

A função novoMapa é uma função auxiliar da "deslizaJogo2" e faz exatamente o mesmo que a função "estendeMapa", mas não permite que sejam geradas linhas cuja velocidade seja superior a 3.

deslizaJogo3

Arguments

:: Int
A função recebe um número inteiro para gerar aleatoriamente uma nova linha.

-> Jogo
A função recebe um jogo (a posição do jogador e um mapa válido).

-> Jogo
A função retornará um novo jogo, com as mesmas coordenadas que o jogador tinha anteriormente mas agora com uma unidade somada ao "y", e o mesmo mapa inicialmente inserido mas com uma nova linha no topo (com as novas propriedade agora definidas) e com a última linha removida.

A função deslizaJogo3 faz exatamente o mesmo que a função "deslizaJogo", mas não permite que sejam geradas linhas cuja velocidade seja superior a 6 e nem relvas com mais do que 4 árvores e nem estradas com mais de 5 carros, e apenas é chamada quando a pontuação do jogador for superior a 200, ou seja serve para que a partir desse valor a dificuldade do jogo aumente.

novoMapa2

Arguments

:: Int
A função recebe um número inteiro para gerar aleatoriamente uma nova linha.

-> Mapa
A função recebe também um mapa válido.

-> Mapa
A função gera um novo mapa válido com a nova linha no topo (cuja velocidade tem de ser inferior a 6).

A função novoMapa2 é uma função auxiliar da "deslizaJogo2" e faz exatamente o mesmo que a função "estendeMapa", mas não permite que sejam geradas linhas cuja velocidade seja superior a 3.

getvelocidade2

Arguments

:: Terreno
A função recebe um tipo de terreno com a sua velocidade correspondente.

-> Int
A função retorna a velocidade correspondente ao terreno que ela recebeu.

A função getvelocidade2 é uma função auxiliar da "novoMapa" e faz exatamente o mesmo que a função "getvelocidade".

reageTempoGloss

Arguments

:: Float
A função recebe um número que corresponde à quantidade de tempo que passa.

-> World
E recebe um valor do tipo World para saber qual o estado atual do jogo.

-> World
A função retorna o novo estado do jogo através do tipo World.

A função reageTempoGloss altera o estado do jogo em funçao do tempo.

main :: IO ()

A função main é a função principal.